+++ CHANGELOG.md
@@ -13,6 +13,10 @@
  `impute_by_recommended`
 * Perform linear regression imputation on all columns with missing values by
  calling linear_regression() without specifying a dependent column.
+ 
+ ### Changed
+
+* Code refactoring and style improvements
 
 ## [0.2](https://github.com/macarro/imputena/releases/tag/v0.2) (2020-04-23)