+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 * Fixed `random_hot_deck_imputation` raising error when specifying a column
  name that contains spaces
+* Fixed `logistic_regression_imputation` raising a ConvergenceWarning in
+ some cases
 * Code refactoring and style improvements
 
 ## [0.2](https://github.com/macarro/imputena/releases/tag/v0.2) (2020-04-23)